+++ CHANGELOG.md
@@ -3,6 +3,7 @@ craftedcore 5.3
 
 - add new lang keys for changing config entries
 - fix duplicate log entries with no internet connection
+- add methods to check caching status (for player profiles)
 
 craftedcore 5.2
 ================